requirements.txt
aio-pika==9.3.0
aiormq==6.7.7
anyio==4.0.0
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
Django==4.2.5
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
pamqp==3.2.1
python-dotenv==1.0.0
python-telegram-bot==20.5
sniffio==1.3.0
sqlparse==0.4.4
yarl==1.9.2
